refactor: myPage를 profileTest로 처리 이후 test용 주석 추가
description:
1) 테스트용 파일명 수정
2) test용 주석 추가
