--- views/components/questionBoardCard copy.ejs
@@ -1,44 +1,0 @@
-<a href="/question/<%= data.qId%>">
-  <div id="cardContainer">
-    <type> <%= data.qType %> | <%= cDate%> </type>
-    <div class="ques">
-      Q.
-      <div class="viewImg">
-        <img
-          src="../../static/svg/eye.svg"
-          alt="조회수"
-          width="24px"
-          class="svg"
-        />
-        <%= data.viewCount%>
-      </div>
-    </div>
-    <div class="ellipsis"><%= data.title%></div>
-    <user>
-      <%= data.uName ? data.uName : "이름없음" %>
-      <!-- 
-        닉네임에 대한 항목은 user 테이블 칼럼명에 존재하여 외래키인 uId를 타고 올라가야 하기 때문에,
-        LEFT OUTER JOIN 수행하는 것을 고려하고 있습니다.
-        추후 JOIN을 진행하게 된다면 말씀 드리겠습니다.  
-      -->
-      <div class="like">
-        <!-- 조회수 추가 여부 -->
-        <img
-          src="../../static/svg/heart.svg"
-          alt="좋아요"
-          width="24px"
-          class="svg"
-        />
-        <%= data.likeCount%>
-        <img
-          src="../../static/svg/message.svg"
-          alt="좋아요"
-          width="24px"
-          class="svg"
-        />
-        <%= data.commentCount ? data.commentCount : 0 %>
-        <!-- 답변갯수 -->
-      </div>
-    </user>
-  </div>
-</a>
